sbd-drones-economics/systems/team1-regulator_operation_devsecops/requirements.txt
aio-pika==9.6.2
aiormq==6.9.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
Authlib==1.6.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.6
dparse==0.6.4
filelock==3.25.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kafka-python==2.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.3
mdurl==0.1.2
multidict==6.7.1
nltk==3.9.4
packaging==26.0
paho-mqtt==2.1.0
pamqp==3.3.0
pluggy==1.6.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-timeout==2.4.0
python-dotenv==1.2.2
python-json-logger==4.0.0
regex==2026.2.28
requests==2.33.0
rich==14.3.3
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
tenacity==9.1.4
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
